finish feature tests for question: view edit create delete

+++ app/controllers/questions_controller.rb
@@ -1,28 +1,38 @@
 # frozen_string_literal: true
 
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: [:index, :show]
   expose :questions, -> { Question.all }
   expose :question
 
   def create
+    self.question = current_user.questions.new(question_params)
     if question.save
-      redirect_to question
+      redirect_to question, notice: "Your question successfully created."
     else
       render :new
     end
   end
 
   def update
-    if question.update(question_params)
-      redirect_to question_path(question)
+    if current_user&.author_of?(question)
+      if question.update(question_params)
+        redirect_to question_path(question)
+      else
+        render :edit
+      end
     else
-      render :edit
+      redirect_to question_path(question), alert: "You are not author of this question."
     end
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user&.author_of?(question)
+      question.destroy
+      redirect_to questions_path, notice: "Your question successfully deleted."
+    else
+      redirect_to question_path, alert: "You are not author of this question."
+    end
   end
 
 private
